+++ backend/src/config/environment.config.ts
@@ -1,4 +1,3 @@
-
 export class Config {
     corsOrigins : string = 'http:localhost:3000';
 }
