+++ backend/src/controllers/search.routes.ts
