+++ backend/src/index.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 import cors from 'cors';
-import { Config } from './config/environment.config';
-import routes from './routes/routes';
+import { Config } from '../config/environment.config';
+import routes from '../routes/routes';
 
 const app = express();
 const config = new Config();
@@ -17,7 +17,6 @@ app.use(cors(
 
 
 app.use(express.json());
-
 app.use('/api', routes)
 
 app.listen(PORT, () => {
